requirements.txt
fastapi
uvicorn
spacy
pytesseract
opencv-python
pillow
pdf2image
pandas
numpy
scikit-learn
python-multipart
pydantic
python-dotenv
tqdm
matplotlib
seaborn
pytest
jupyter
transformers
torch